refactor: 패지키 구조 정리
