--- src/main/java/until/the/eternity/auction/repository/AuctionHistoryRepositoryCustom.java
@@ -1,10 +1,0 @@
-package until.the.eternity.auction.repository;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import until.the.eternity.auction.domain.dto.internal.request.AuctionHistorySearchRequest;
-import until.the.eternity.auction.domain.model.AuctionHistory;
-
-public interface AuctionHistoryRepositoryCustom {
-    Page<AuctionHistory> search(AuctionHistorySearchRequest condition, Pageable pageable);
-}
