+++ src/main/java/until/the/eternity/auctionhistory/domain/dto/external/OpenApiItemOptionResponse.java
@@ -1,0 +1,10 @@
+package until.the.eternity.auctionhistory.domain.dto.external;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record OpenApiItemOptionResponse(
+        @JsonProperty("option_type") String optionType,
+        @JsonProperty("option_sub_type") String optionSubType,
+        @JsonProperty("option_value") String optionValue,
+        @JsonProperty("option_value2") String optionValue2,
+        @JsonProperty("option_desc") String optionDesc) {}
