+++ src/main/java/until/the/eternity/auctionhistory/domain/dto/internal/response/ItemOptionResponse.java
@@ -1,0 +1,9 @@
+package until.the.eternity.auctionhistory.domain.dto.internal.response;
+
+public record ItemOptionResponse(
+        Long id,
+        String optionType,
+        String optionSubType,
+        String optionValue,
+        String optionValue2,
+        String optionDesc) {}
