+++ src/main/java/until/the/eternity/auctionitem/domain/entity/AuctionItem.java
@@ -1,0 +1,39 @@
+package until.the.eternity.auctionitem.domain.entity;
+
+import jakarta.persistence.*;
+import java.time.LocalDateTime;
+import java.util.List;
+import lombok.*;
+import until.the.eternity.itemoption.domain.entity.ItemOption;
+
+@Entity
+@Table(name = "auction_item")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class AuctionItem {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "item_name", nullable = false)
+    private String itemName;
+
+    @Column(name = "item_display_name", nullable = false)
+    private String itemDisplayName;
+
+    @Column(name = "item_count", nullable = false)
+    private Long itemCount;
+
+    @Column(name = "auction_price_per_unit", nullable = false)
+    private Long auctionPricePerUnit;
+
+    @Column(name = "date_auction_expire", nullable = false)
+    private LocalDateTime dateAuctionExpire;
+
+    @OneToMany(mappedBy = "auctionItem", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<ItemOption> itemOptions;
+}
